requirements.txt
streamlit==1.45.0
langchain==0.3.25
langchain-groq==0.3.2
langchain-huggingface==0.1.2
langchain-community==0.3.24
langchain-classic==0.1.0
faiss-cpu==1.11.0
pypdf==5.4.0
python-dotenv==1.1.0
sentence-transformers==4.1.0
rank-bm25==0.2.2